Moved Rpc class tests to folders
