+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpLib.UnitTests.csproj
@@ -46,19 +46,19 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="IceWarpObjects\Rpc\Classes\BaseTest.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleAllCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleConditions_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleDNSBLCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleHasAttachmentCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleIsSizeCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleIsSpamCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleLocalTimeCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRulePriorityCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleSenderRecipientCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleSMTPAuthCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleSomeWordsCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleSpamScoreCondition_Test.cs" />
-    <Compile Include="IceWarpObjects\Rpc\Classes\TRuleTrustedSessionCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleAllCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleConditions_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleDNSBLCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleHasAttachmentCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleIsSizeCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleIsSpamCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleLocalTimeCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRulePriorityCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleSenderRecipientCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleSMTPAuthCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleSomeWordsCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleSpamScoreCondition_Test.cs" />
+    <Compile Include="IceWarpObjects\Rpc\Classes\Rule\TRuleTrustedSessionCondition_Test.cs" />
     <Compile Include="IceWarpRpc\Requests\Account\AccountRequestTests.cs" />
     <Compile Include="IceWarpRpc\Requests\Domain\DomainRequestTests.cs" />
     <Compile Include="IceWarpRpc\Requests\ErrorResponseTests.cs" />
@@ -308,6 +308,11 @@
       <Project>{87007c89-ecde-4130-aa26-5af14b251379}</Project>
       <Name>IceWarpLib.Rpc</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="IceWarpObjects\Rpc\Classes\Account\" />
+    <Folder Include="IceWarpObjects\Rpc\Classes\Domain\" />
+    <Folder Include="IceWarpObjects\Rpc\Classes\Property\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
