+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/TRuleHasAttachmentCondition_Test.cs
@@ -1,16 +1,14 @@
-﻿using IceWarpLib.Objects.Helpers;
-using IceWarpLib.Objects.Rpc.Classes;
-using IceWarpLib.Objects.Rpc.Classes.Rule;
+﻿using IceWarpLib.Objects.Rpc.Classes.Rule;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
 {
-    public class TRuleTrustedSessionCondition_Test : BaseTest
+    public class TRuleHasAttachmentCondition_Test : BaseTest
     {
         private string _xml = @"
 <custom xmlns=""admin:iq:rpc"">
-    <classname>truletrustedsessioncondition</classname>
+    <classname>trulehasattachmentcondition</classname>
     <conditiontype>0</conditiontype>
     <operatorand>0</operatorand>
     <logicalnot>0</logicalnot>
@@ -19,18 +17,18 @@ namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes
 </custom>".TrimStart();
 
         [Test]
-        public void TRuleTrustedSessionCondition()
+        public void TRuleHasAttachmentCondition()
         {
-            var testClass = new TRuleTrustedSessionCondition();
+            var testClass = new TRuleHasAttachmentCondition();
 
             var testXml = ToFormattedXml(testClass);
             Assert.AreEqual(_xml, testXml);
         }
 
         [Test]
-        public void TRuleTrustedSessionCondition_BuildXmlElement()
+        public void TRuleHasAttachmentCondition_BuildXmlElement()
         {
-            var testClass = new TRuleTrustedSessionCondition(GetXmlNode(_xml));
+            var testClass = new TRuleHasAttachmentCondition(GetXmlNode(_xml));
 
             Assert.AreEqual(TRuleConditionType.None, testClass.ConditionType);
         }
