+++ IceWarpLib/IceWarpLib.UnitTests/IceWarpObjects/Rpc/Classes/Rule/TRuleSMTPAuthCondition_Test.cs
@@ -1,37 +1,34 @@
-﻿using IceWarpLib.Objects.Helpers;
-using IceWarpLib.Objects.Rpc.Classes;
-using IceWarpLib.Objects.Rpc.Classes.Rule;
+﻿using IceWarpLib.Objects.Rpc.Classes.Rule;
 using IceWarpLib.Objects.Rpc.Enums;
 using NUnit.Framework;
 
-namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes
+namespace IceWarpLib.UnitTests.IceWarpObjects.Rpc.Classes.Rule
 {
-    public class TRulePriorityCondition_Test : BaseTest
+    public class TRuleSMTPAuthCondition_Test : BaseTest
     {
         private string _xml = @"
 <custom xmlns=""admin:iq:rpc"">
-    <classname>truleprioritycondition</classname>
+    <classname>trulesmtpauthcondition</classname>
     <conditiontype>0</conditiontype>
     <operatorand>0</operatorand>
     <logicalnot>0</logicalnot>
     <bracketsleft>0</bracketsleft>
     <bracketsright>0</bracketsright>
-    <priority>0</priority>
 </custom>".TrimStart();
 
         [Test]
-        public void TRulePriorityCondition()
+        public void TRuleSMTPAuthCondition()
         {
-            var testClass = new TRulePriorityCondition();
+            var testClass = new TRuleSMTPAuthCondition();
 
             var testXml = ToFormattedXml(testClass);
             Assert.AreEqual(_xml, testXml);
         }
 
         [Test]
-        public void TRulePriorityCondition_BuildXmlElement()
+        public void TRuleSMTPAuthCondition_BuildXmlElement()
         {
-            var testClass = new TRulePriorityCondition(GetXmlNode(_xml));
+            var testClass = new TRuleSMTPAuthCondition(GetXmlNode(_xml));
 
             Assert.AreEqual(TRuleConditionType.None, testClass.ConditionType);
         }
